TestDB download PreprocessedData getGenerated

+++ ArCo-release/rdfizer/src/test/java/arco/rdfizer/TestDB.java
@@ -1,9 +1,11 @@
 package arco.rdfizer;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 import java.io.File;
 import java.io.IOException;
+import java.lang.reflect.Field;
 
 import org.junit.Test;
 
@@ -33,10 +35,43 @@ public class TestDB {
 			assertEquals("http://www.sigecweb.beniculturali.it/images/fullsize/ICCD1060938/ICCD13814908_F1_esterno.jpg",
 					pd.getFtan2URL().get("3969"));
 
+			new File(dbFile).delete();
+			resetInstance();
 		} catch (IOException e) {
 			e.printStackTrace();
+		} catch (NoSuchFieldException e) {
+			e.printStackTrace();
+		} catch (SecurityException e) {
+			e.printStackTrace();
+		} catch (IllegalArgumentException e) {
+			e.printStackTrace();
+		} catch (IllegalAccessException e) {
+			e.printStackTrace();
 		}
-		new File(dbFile).delete();
+	}
+
+	private static void resetInstance()
+			throws NoSuchFieldException, SecurityException, IllegalArgumentException, IllegalAccessException {
+		Field privateStringField = PreprocessedData.class.getDeclaredField("instance");
+		privateStringField.setAccessible(true);
+		privateStringField.set(PreprocessedData.class, null);
+	}
+
+	@Test
+	public void testDownload() {
+		if (System.getenv("USER").equals("lgu")) {
+			String dbFile = "test_db";
+			System.out.println("Test download");
+			new File(dbFile).delete();
+			PreprocessedData pd = PreprocessedData.getInstance(true, dbFile, false);
+			assertTrue(pd.getGenerated() > 0);
+			new File(dbFile).delete();
+			try {
+				resetInstance();
+			} catch (NoSuchFieldException | SecurityException | IllegalArgumentException | IllegalAccessException e) {
+				e.printStackTrace();
+			}
+		}
 	}
 
 }
